Using Enum(n) rather than deprecated Enum(UInt(), n).

+++ src/main/scala/examples/Parity.scala
@@ -9,7 +9,7 @@ class Parity extends Module {
     val in  = Input(Bool())
     val out = Output(Bool())
   })
-  val s_even :: s_odd :: Nil = Enum(UInt(), 2)
+  val s_even :: s_odd :: Nil = Enum(2)
   val state  = Reg(init=s_even)
   when (io.in) {
     when (state === s_even) { state := s_odd  }
