+++ src/main/scala/solutions/VendingMachineSwitch.scala
@@ -10,9 +10,9 @@ class VendingMachineSwitch extends Module {
     val dime   = Input(Bool())
     val valid  = Output(Bool())
   })
-  val s_idle :: s_5 :: s_10 :: s_15 :: s_ok :: Nil = Enum(UInt(), 5)
+  val s_idle :: s_5 :: s_10 :: s_15 :: s_ok :: Nil = Enum(5)
   val state = Reg(init = s_idle)
-  
+
   switch (state) {
     is (s_idle) {
       when (io.nickel) { state := s_5 }
